Create MonthCharts component to reduce copying
